webapi/webapi-serviceworkers-w3c-tests/serviceworkers/clients-claim.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2015 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>clients.claim()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="https://slightlyoff.github.io/ServiceWorker/spec/service_worker/index.html#clients-claim-method">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="w3c/service-workers/resources/test-helpers.js"></script>
<div id="log"></div>
<script>

promise_test(function(t) {
  var scope1 = "resources/";
  var scope2 = "resources/blank-iframe.html";
  var script1 = "resources/empty-worker.js";
  var script2 = "resources/clients-claim-worker.js";
  var frame;
  var worker;
  service_worker_unregister_and_register(t, script1, scope1)
  .then(function(r) {
    return wait_for_state(t, r.installing, "activated");
  })
  .then(function () {
    return with_iframe(scope2);
  })
  .then(function(fr) {
    frame = fr;
    return navigator.serviceWorker.register(script2, {scope: scope2});
  })
  .then(function(r) {
    registration = r;
    var frameWorker = frame.contentWindow.navigator.serviceWorker;
    var controllerChanged = new Promise(function(resolve) {
      frameWorker.oncontrollerchange = resolve;
    });
    var controllerMessage = new Promise(function(resolve) {
      frameWorker.onmessage = t.step_func(function(evt) {
        assert_true(evt.data, "clients.claim() return undefined");
        resolve();
      });
    });
    return Promise.all([controllerChanged, controllerMessage]);
  })
  .then(function() {
    var url = normalizeURL(script2);
    assert_equals(
      frame.contentWindow.navigator.serviceWorker.controller.scriptURL,
      url,
      "Frame controller scriptURL should be changed to script2");
    frame.remove();
    service_worker_unregister_and_done(t, scope);
  })
  .catch(unreached_rejection(t));
}, "Check that clients.claim() return undefined.");

</script>
